Refactor app controller
