+++ src/followers/followers.module.ts
@@ -1,8 +1,11 @@
 import { Module } from '@nestjs/common';
 import { FollowersService } from './followers.service';
 import { FollowersController } from './followers.controller';
+import { UsersModule } from 'src/users/users.module';
+import { MoviesModule } from 'src/movies/movies.module';
 
 @Module({
+  imports: [UsersModule, MoviesModule],
   providers: [FollowersService],
   controllers: [FollowersController],
   exports: [FollowersService],
