+++ views/follower.ejs
@@ -83,7 +83,7 @@
                 src="<%= ratedMovie.movie.poster %>"
                 alt="<%= ratedMovie.movie.title %> poster"
               />
-              <a href="/movie/<%= ratedMovie.movie.imdbID %>">
+              <a href="/movies/<%= ratedMovie.movie.imdbID %>">
                 <h3><%= ratedMovie.movie.title %></h3>
               </a>
               <div>
