+++ views/login.ejs
@@ -19,7 +19,7 @@
         </div>
         <div></div>
         <div class="auth">
-          <a href="/register">Register</a>
+          <a href="/users/register">Register</a>
         </div>
       </nav>
     </header>
