+++ views/movie.ejs
@@ -32,7 +32,7 @@
 
     <main class="main">
       <div class="box">
-        <a href="../" class="btn-back">x</a>
+        <a href="/movies" class="btn-back">x</a>
         <div class="details">
           <header>
             <img
@@ -75,7 +75,7 @@
         </div>
       </div>
       <div class="box">
-        <a href="../" class="btn-back">x</a>
+        <a href="/movies" class="btn-back">x</a>
         <div class="summary">
           <h2>Ratings from followers</h2>
           <div>
@@ -92,7 +92,7 @@
                 width="40px"
                 alt="<%= follower.followerName %>"
               />
-              <a href="/follower/<%= follower.followerId %>">
+              <a href="/followers/<%= follower.followerId %>">
                 <h3><%= follower.followerName %></h3>
               </a>
               <div>
