+++ views/movies.ejs
@@ -56,7 +56,7 @@
                 src="<%= ratedMovie.movie.poster %>"
                 alt="<%= ratedMovie.movie.title %> poster"
               />
-              <a href="/movie/<%= ratedMovie.movie.imdbID %>">
+              <a href="/movies/<%= ratedMovie.movie.imdbID %>">
                 <h3><%= ratedMovie.movie.title %></h3>
               </a>
               <div>
@@ -116,7 +116,7 @@
           const movieItem = document.createElement('li');
           movieItem.innerHTML = `
             <img src="${movie.Poster}" alt="${movie.Title} poster">
-            <a href='/movie/${movie.imdbID}'>
+            <a href='/movies/${movie.imdbID}'>
                 <h3>${movie.Title}</h3>
             </a>
             <div>
