+++ views/register.ejs
@@ -19,7 +19,7 @@
         </div>
         <div></div>
         <div class="auth">
-          <a href="/login">Login</a>
+          <a href="/users/login">Login</a>
         </div>
       </nav>
     </header>
